Create Dekad_arran_xl-netcdf.py
This conversion is done over Northern Ghana. The script picks station data in excel format and reconstructs a netcdf formatted data from it.
The second part of the script selects specific longitudes and latitudes (stations) from a grid data . This script can be modified and replicated for other data files and regions.